requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
torch>=2.2.0
streamlit>=1.41.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
requests>=2.32.0
kaggle>=1.6.17
python-dotenv>=1.0.1
joblib>=1.4.2
networkx>=3.4.2
PyYAML>=6.0.2
pytest>=8.3.0
httpx>=0.27.0
python-multipart>=0.0.9
